backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.13.0
httpx>=0.25.0
requests>=2.31.0  # For Workload Identity ID token fetching (FR#080)
google-genai>=0.3.0  # Official Google GenAI Python SDK for Gemini model interaction (FR#090)
google-auth>=2.20.0  # For Workload Identity ID token verification (FR#335)
google-auth-oauthlib>=1.0.0  # For WI token handling (FR#335)
# TODO: Add google-adk when available (Google Agent Development Kit)
# google-adk>=X.X.X

